[starnix] Don't hardcode iface name in `SIOCGIFINDEX`
Before this change, `ioctl(_, SIOCGIFINDEX, _)` returned a value of of
`1` if the requested interface name was `sta-iface-name`. The value `1`
is usually assigned to the loopback interface has a different name and
the name `sta-iface-name` may not be present on all systems.

The changes implements `ioctl(_, SIOCGIFINDEX, _)` (for inet sockets)
by requesting the interface index for the interface name from the
netstack.

Bug: 128604, 129059
Bug: b/286855878
Change-Id: I4c8cf3eaf8a011ba9f8b970a3c330bb87e9a7a89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870847
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>