[profiler] Drop dependency on inspector
This cl changes to manually invoke elfsearch instead of going through
the inspector api which awkwardly wrote out to a FILE. By manually
collecting the information we can now directly write to the socket and
don't need the temp file.

This also helps us have the symbolization information in a structured
format for if we change the output to be structured instead of in
symbolizer markup.

Change-Id: Ib0aecbf53de19687e358936a98b7289cb7b71f79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862881
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>