[boot-shim] Bootstrap '/chosen'
For devicetree item, the early stage where no allocations are allowed,
require use to bootstrap memory layout, cmdline and the
uart(diagnostics).

Thist CL, introduces a stub for the non boostrapping phase, and
an implementation for the bootstrapping phase.

In order to unit test this item in all platforms, the
uart::all::KernelDriver was slightly modifed in order to not assume
uart::all::Drivers always. It was left as the default option.

Introduce MaybeCreate for devicetree binding

Devicetree device nodes specify a list of strings that represent devices
which they are compatible with. The property name is 'compatible'.

This work is required to instantiate the uart driver from the devicetree
blob.

Bug: 78675
Test: fx test uart-unittests
Change-Id: Ibf9b559290457c5cfed15559008bc93bfc515070
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860576
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>