[trace] Manage rust trace contexts via RAII wrapper
Some specialized trace producers may want to be able to avoid the
computation to compute trace arguments, events, etc when tracing is not
enabled. For this purpose, it is more natural to try acquiring a
context and use that context to write the record after said computation,
rather than first checking if tracing is enabled for the category and
then later obtaining the context which would result in a second check
(and may also give a different result from the first check).

Bug: b/271467732

Change-Id: Ibdefd73cd59868fe5c1f380ba55b826de67bcef6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869159
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>