[fuchsia-component] Make ServiceFs.add_entry_at public
Let clients add any kind of directory entry.

The existing add_* methods are not always sufficient, e.g. it's not
possible to have `add_remote` forward to the exposed directory of
a component that is recreated at runtime because it only accepts
a single proxy.

Change-Id: If511c190b3864f7692f9cf10f823fbc03a17045d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871303
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>