[boot-shim] Introduce DevicetreeMemoryItem
This device item is in charge of digesting memory items within a
devicetree. That is, memory nodes, reserved-memory nodes and memreserve
ranges.

Test include individual case parsing, and a mixed case that includes
checking the verified ZBI image.

This Item is meant to be used as a bootstrapping item, that will allow
general matchers to perform allocations.

Bug: 78675
Test: fx test boot-shim-unittests
Change-Id: I7b53ce564111053a5f9306e9f5d71c976037bf10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858118
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>