[build] Fix rust_proto_library() hermeticity
The rust_proto_library() template invokes the `prost-build`
Rust program that does not generate depfiles at all, so this
CL wraps the call with a wrapper script that can compute the
list of transitive protobuf imports.

This allows generating a proper depfile, and verifying that
the `packages` value passed to rust_proto_library() is also
correct. If not, a human-friendly error message is printed
explaining the issue.

Fixed: 127778

Change-Id: I9643dac400ad01c50757e6d58a41c1ae1035fecc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873217
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>