[starnix] Add a minimal tracefs mount type
This adds a tracefs file system that contains only a single node:
trace_marker. Unlike tracefs on a Linux system, this implementation does
not maintain a ring buffer of trace events, so it does not expose trace
events via readable files.

When data is written to the trace_marker file, if the data is in the
format used by atrace, it will be translated to a Fuchsia trace event
and written to the Fuchsia tracing system.

Bug: b/271467732
Change-Id: I9cec17408bca254a258afaa4c3ae852361b3b984
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869977
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>