[fxblob] Add an as_blob option to Volumes.Create
Volumes.Create should not rely on the name of the volume (i.e. checking
if the name is "blob") to determine how to mount the volume. Strings
are a weak API. Instead, Volumes.Create now takes in MountOptions which
contains an as_blob option. When as_blob is set, the volume will be
mounted as a blob filesystem.

Fixed: 126745
Change-Id: I9f8f33b5ce174d8716b3227163c15504977f5455
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870176
Commit-Queue: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>