[check-licenses] Enable license-classifier.
Previously, check-licenses used simple regex pattern matching to
detect licenses in a text file. This works, but setting up and
maintaining the regexp patterns so they work for every project is
painful and error prone.

https://github.com/google/licenseclassifier is a library that
can generate a similarity score between a given blob of text and
a known license text library. In practice it is very accurate,
fast, concurrency safe, and the library of license texts is
significantly easier to generate and maintain.

This CL flips the switch from the old strategy to the new one.
* Remove references of the "license" package across the project.
 -> that package implemented the regexp solution.
* Adjust allowlists to temporarily allow a handful of new issues

Fixed: 113024, 111682, 102068
Change-Id: Ie448f0033bb6d17b3982a5c2a94feb19623e8b04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868757
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>