[fuchsia-controller] Support isolation.
This adds support for an isolate_dir object to be passed to the fuchsia
controller context. When this isolate dir goes out of scope, the
directory it points to will be removed to clean up the FFX daemon (the
ffx daemon will shut itself down once the socket it's pointing to is
deleted).

Fixed: 129193
Multiply: fuchsia_controller_e2e_test
Change-Id: I2216f2c6750b1d76063b6848433e6630bb1590ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872267
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>