[starnix] Avoid an unwrap when upgrading `Weak`
When a `ProcDirectory` is created, we have access to an `Arc` but we
pass a downgraded `Weak` which we later try to upgrade. This requires
an `unwrap` which we can avoid by passing the `Arc` as far as possible.

Change-Id: I58fe9755f5db6c48ef894b7067415cbdb52786c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873920
Reviewed-by: Theodore Dubois <tbodt@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>