[unwinder] Report pc_is_return_address explicitly
In the past pc_is_return_address is inferred implicitly from the index
of the frame or the trust level of the frame because asynchronously
created frames such as signal frames are not supported on Fuchsia. With
starnix being able to unwind into restricted mode, the first frame in
the restricted mode should also be marked as :pc (precise code location)
rather than :ra.

Change-Id: I11593c9ad9a9e8bd4ff20dd11ca91eb0597c4a77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874696
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>