[bt][hfp-ag] Set IO coding format correctly.
The BT Spec says that the IO Coding Format should be Transparent if and
only if the Transmit_Coding_format and Receive_Coding_Format are
Transparent.  Update the formats that we request to match with in-band
audio encoding.

Bug: 122491
Test: fx build and test with inband enabled
Change-Id: Id395c4c50e2f075fcf1a640ab38111348dfbcd1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866576
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>