[tracing] Add Additional Help for Connection Errors
Added a small blurb with some possible diagnosing steps and a pointer to
the tracing documentation when tracing fails to connect to trace
manager.

When it fails to connect it now prints out this message:

```
Error: ffx trace was unable to connect to trace_manager on the device.

Note that tracing is available for eng and core products, but not user or userdebug.
To fix general connection issues, you could also try:

$ ffx doctor

For a tutorial on getting started with tracing, visit:
https://fuchsia.dev/fuchsia-src/development/sdk/ffx/record-traces
```

Change-Id: I9145b874f9ff5d6f1faed76741d5dc2a4224e019
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/875718
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>