[netstack] Convert interface auth grant to proof
Add a helper fn to convert from `GrantForInterfaceAuthorization`
to `ProofOfInterfaceAuthorization` to be used as part of authorizing
routes to be added that use a particular interface.

Jeff and I discovered that it is not possible to create an EventPair
with an `INSPECT` right, and that it is not necessary to request
the basic info about an `EventPair`. Therefore, it has been removed
from the FIDL definition and now `ProofOfInterfaceAuthorization`
does not need any rights.

It is not possible to specify `NONE` as the Rights on the token, so a
bug has been created to update it when possible.

Change-Id: Ia3dacbc295cb9b4802e68703b557e10692ccd7a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874337
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Antonio Linhart <antoniolinhart@google.com>
API-Review: Ghanan Gowripalan <ghanan@google.com>