[pbv2] introduce LoadedProductBundle to track where product bundles are loaded from.
This CL adds a new struct which wraps the ProductBundle and allows us to
track where the product bundle was loaded from even if we used the
configuration domains to load it. This will be used for the emulator
code paths to let the developer know what product bundle is being
loaded. The users of the load_product_bundle method have been updated to
call into() which immediately converts the object back to the
ProductBundle enum. In the future these call sites can be updated to use
the LoadedProductBundle.

In the future the ProductBundle will be converted from an enum to
a Trait so that we do not need to introduce this extra layer.

Change-Id: I3b3909709c131dc015e3c5a52faee47a9695ff4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872789
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>