[net/mdns] honor cache flush flag for subscribed host names
This CL fixes an issue in which the 'host name requestor' is not
honoring the cache flush bit in address records (A/AAAA). This
bit should cause previously-seen addresses for the target in
question to be forgotten.

This CL also addresses problems with the way addresses were handled
previously. Because cache flush behavior was not understood when
this agent was originally written, there was an attempt to use message
boundaries to determine when to purge old addresses. That logic has
been replaced with proper cache flush handling.

Test: new unit test added

Bug: 128612

Change-Id: Ifd41d6d6a2dbfcb73b9cec8be9f4a03e194703a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874541
Reviewed-by: Michael Spang <spang@google.com>
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>