[netlink] Netlink returns negative Errnos
Ensure that the Errnos returned by the netlink crate are negated, as
specified on the Netlink man page.

While here, refactor the error creation path in the to disambiguate
between `Ack` and `Error`. This change is more aligned with the upstream
`netlink-packet-core` crate.

Fixed: b/288922671

Change-Id: I64c9754921159f694eaf83a5b2daf21ea39df436
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/876597
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>