[driver][sdk][testing] runtime rework
This change introduces a change into how dispatchers are created and run
in driver unit tests. It creates a new DriverRuntime class that is in
charge of creating the foreground and background dispatchers, as well as
running them. It provides many of the same APIs that async::Loop does
like Run and Quit.

Bug: 127188

Change-Id: Iba846f8af0573f1fc84324b68c048ce6cc66b902
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871508
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>