[netstack] Specify TRANSFER rights on `Proof`
In another CL, we found out that the TRANSFER right is required to send
the EventPair across a FIDL channel at all, even if it is not explicitly
being duplicated/replaced.

While not necessary for functionality, the FIDL was updated to reflect
the lowest level of rights necessary to follow the guidelines for
specifying handle rights: see RFC-0028: Handle rights for more info.

Fix: 129584
Change-Id: Ia71cb6d96990fbb658eab67952c41a57f46e05b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877516
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
API-Review: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>