[archivist] Remove Trie and UniqueKey in repositories
This refactor simplifies the follow-up which gets rid of
ComponentIdentifier (in favor of ExtendedMoniker) and simplifies a
lot of code.

As part of this change we can now delete the "UniqueKey" which was
needed only for CFv1 purposes.

The Trie was introduced at the time for two reasons:

1) Components in the v1 topology could share the same moniker and we'd
   distinguish them by the instance id. We no longer have v1 so we can
   stop supporting this.

2) Do not repeat the key multiple times. Since then, we now have string
   interning via FlyStr so we no longer need that. We can just use the
   ComponentIdentity as the key.

Change-Id: Idc75877cc99e6ce6ba2b401d3bccd539a34c4597
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877437
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>