[audio] Fix audio_driver_tests use-after-free
Fix a possible usage after free when a device is detected after
an initial device detection is completed. Related to this, all
device-related timeouts are eliminated, so we rely on the test
execution framework to timeout and kill the test suite if the
device interactions time out.

The DetectDevices method was incorrectly using stack variables
in a callback that could be invoked after the method completed.
DetectDevices is not designed to keep discovering devices after
it completes. However, device_watchers_ are stored to keep the
channel used to connect to the devices, which can lead to
DeviceWatcher callbacks after DetectDevices completes.

In this CL we make sure any subsequent callbacks are treated
as errors. We further safeguard these variables against misuse
in the future by moving them to static storage.

A bug is fixed, related to the number of virtual_audio
instances that are added. This actually was the underlying
issue that uncovered the use-after-free.

Fixed: 129582

Test: audio_driver_tests in asan build
Change-Id: Iba7f81826c17811258217ba1d68aa80b7e509812
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/877677
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>