[security] Use real fshost in pkg_test
This is done to support the fxfs+fxblob filesystem configuration, in
addition to the fvm+blobfs configuration that currently runs.

This switches pkg_test to use the fshost_component_and_config template
to create fshost instead of using a fake.

It will also switch assembled_system to generate a fxfs image instead of
fvm in the fxfs+fxblob configuration. Either image will be correctly
interpreted by fshost when loaded into a VMO and handed to
RamdiskClient, which adds the images to the isolated dev tree.

Configuration and routing of dependencies for fshost follows examples
from the fshost integration tests at //src/storage/fshost/integration.

Bug: 125808
Test: fx test -o //src/security/tests/pkg_test/tests
Change-Id: Iabc98c310fb9bd7657817345db7de3fc20af6eec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869476
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>