[fxfs] Build a sparse image in assembly
Generates an image `fxfs.sparse.blk` which is in the Android Sparse
format.  This image is provided in the flash manifest, which is already
handled by the fastboot protocol in a more efficient manner.

The sparse image is significantly smaller than the raw image, since all
padding can be represented in a few bytes.

This change includes substantial refactoring of the `sparse` library, as
well as adding the `unsparse` method (which is currently only used for
testing).

The equivalence of the sparse image to the raw image was manually
verified with the `simg2img` tool.  The expanded fxfs.sparse.blk is
identical to fxfs.blk.

Change-Id: I22d4d813094f790d0fb241fd5d16cc630e84efd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873758
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>