[memory_sampler] (Poisson) sampling in bytes between samples.
With the current instrumentation strategy, reporting every single
allocation and deallocation is prohibitively expensive on
moderately-to-large sized processes: the FIDL channel used to
communicate data with the profiler service fills up too fast in the
absence of backpressure, causing the connection to eventually be
severed. To mitigate this, this instrumentation randomly samples the
allocations to be reported, hence providing a way to reduce the
instrumentation overhead at the cost of accuracy.

We use a sampling strategy similar to that of Chromium's and
tcmalloc's sampling heap profilers: Poisson sampling with a constant
average sampling interval, in number of bytes allocated between
samples.

This initial support embeds a constant interval; we intend to
introduce configurability in a subsequent commit.

Change-Id: I70ba02e3034a5490bee225ce1d67f692c7cbfb5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864019
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Fabio D'Urso <fdurso@google.com>